code polishing to remove warnings at compilation

+++ modules/ustk_core/include/visp3/ustk_core/usImagePreScan2D.h
@@ -90,6 +90,9 @@ private:
   double m_axialResolution;
 };
 
+/**
+* Basic constructor, all settings set to default. For double data.
+*/
 template<class T>
 usImagePreScan2D<T>::usImagePreScan2D() : vpImage<T>(), usImageSettings()
 {
@@ -116,7 +119,7 @@ usImagePreScan2D<T>::usImagePreScan2D(unsigned int AN, unsigned int LN) : vpImag
 * @param isConvex Boolean to specify if the probe used was convex(true) or linear(false).
 */
 template<class T>
-usImagePreScan2D<T>::usImagePreScan2D(unsigned int AN, unsigned int LN, float probeRadius, float scanLinePitch, bool isConvex) :
+usImagePreScan2D<T>::usImagePreScan2D(unsigned int AN, unsigned int LN, double probeRadius, double scanLinePitch, bool isConvex) :
     vpImage<T>(AN, LN), usImageSettings(probeRadius, scanLinePitch, isConvex)
 {
